refactor: separated ViewUtils to Color and Text utils
